merge

+++ lib/features/adary/presentation/widgets/health/health_item.dart
@@ -3,6 +3,7 @@ import 'package:adary/core/conts/icons.dart';
 import 'package:adary/core/share/widgets/btn_icon.dart';
 import 'package:adary/core/share/widgets/container_btns.dart';
 import 'package:adary/core/share/widgets/expansion_widget.dart';
+import 'package:adary/core/utils/app_utils.dart';
 import 'package:adary/features/adary/data/models/health_model.dart';
 import 'package:adary/features/adary/domain/entities/delete_entity.dart';
 import 'package:adary/features/adary/presentation/bloc/health/health_bloc.dart';
@@ -104,43 +105,51 @@ class HealthItem extends StatelessWidget {
               content: Row(
                 mainAxisAlignment: MainAxisAlignment.spaceAround,
                 children: [
-                  BtnIcon(
-                      label: 'delete'.tr(),
-                      icon: AppIcon.rash,
-                      onTap: () {
-                        AwesomeDialog(
+                  if (AppUtils.permissions.isNotEmpty &&
+                          AppUtils.permissions.any(
+                              (p) => p.contains("api/notes/health/delete/")) ||
+                      AppUtils.permissions.isEmpty)
+                    BtnIcon(
+                        label: 'delete'.tr(),
+                        icon: AppIcon.rash,
+                        onTap: () {
+                          AwesomeDialog(
+                              context: context,
+                              dialogType: DialogType.warning,
+                              titleTextStyle: const TextStyle(
+                                  color: Colors.red,
+                                  fontSize: 20,
+                                  fontWeight: FontWeight.bold),
+                              title: 'delete_health'.tr(),
+                              desc: 'delete_health_des'.tr(),
+                              btnCancelText: 'no'.tr(),
+                              btnOkText: 'delete'.tr(),
+                              btnCancelOnPress: () {},
+                              btnOkOnPress: () {
+                                BaseBloc.get<HealthBloc>(context).add(
+                                    DeleteHealthEvent(
+                                        baseEnity: DeleteEntity(
+                                            id: visitModel.id ?? 0)));
+                              }).show();
+                        }),
+                  if (AppUtils.permissions.isNotEmpty &&
+                          AppUtils.permissions.any(
+                              (p) => p.contains("api/notes/health/update/")) ||
+                      AppUtils.permissions.isEmpty)
+                    BtnIcon(
+                        label: 'edit'.tr(),
+                        icon: AppIcon.edit,
+                        onTap: () {
+                          showModalBottomSheet(
+                            isScrollControlled: true,
                             context: context,
-                            dialogType: DialogType.warning,
-                            titleTextStyle: const TextStyle(
-                                color: Colors.red,
-                                fontSize: 20,
-                                fontWeight: FontWeight.bold),
-                            title: 'delete_health'.tr(),
-                            desc: 'delete_health_des'.tr(),
-                            btnCancelText: 'no'.tr(),
-                            btnOkText: 'delete'.tr(),
-                            btnCancelOnPress: () {},
-                            btnOkOnPress: () {
-                              BaseBloc.get<HealthBloc>(context).add(
-                                  DeleteHealthEvent(
-                                      baseEnity: DeleteEntity(
-                                          id: visitModel.id ?? 0)));
-                            }).show();
-                      }),
-                  BtnIcon(
-                      label: 'edit'.tr(),
-                      icon: AppIcon.edit,
-                      onTap: () {
-                        showModalBottomSheet(
-                          isScrollControlled: true,
-                          context: context,
-                          builder: (context) {
-                            return AddHeathPage(
-                              healthCondition: visitModel,
-                            );
-                          },
-                        );
-                      })
+                            builder: (context) {
+                              return AddHeathPage(
+                                healthCondition: visitModel,
+                              );
+                            },
+                          );
+                        })
                 ],
               ),
             )
